requirements.txt
autopep8==1.5.4
boto3==1.17.57
botocore==1.20.57
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
des==1.0.5
filesplit==3.0.2
idna==2.10
jmespath==0.10.0
Naked==0.1.31
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
shellescape==3.8.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
